split ReportController into week and month controller
